backend/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv
Werkzeug
